test: add build->run dep regression test in TestConcretizeSeparately
This introduces a regression test to ensure that the concretizer fails
when trying to solve a spec where we have a build tool in two versions and
the "lower" (in the DAG) version being also a run-dependency. If this
would be allowed, we could end up having multiple paths to different tool
versions/variants in the build env.

Change-Id: I050efc9471d89580f5546a1c56ab419891bcc792