feat: describe python dependency in package
Previously, the python was added as a dependency during the
concretization (as a command line argument) in yashchiki. As a result,
python was not a direct dependency of the "visionary packages" such as
visionary-dls or visionary-wafer but an implicit dependency.
Furthermore, a global python version was enforced for all packages in
the same container (even across different apps).

By defining the dependency in the packages directly, the dependency is
explicit and we can have different python versions across different
apps.

Depends-On: 24335

Change-Id: Ibf6eb55e9547ac556fa763c0de2c291960488c2c